requirements.txt
alembic==1.5.8
aniso8601==9.0.1
appdirs==1.4.3
attrs==20.3.0
bcrypt==3.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.1.0
Flask-Migrate==2.7.0
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lockfile==0.12.2
Mako==1.1.4
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.16.0
python-editor==1.0.4
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.15.0
SQLAlchemy==1.4.3
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
